Moved BECMGChange to trend package
